requirements.txt
# tensorrt~=10.0.1
pycuda~=2024.1

onnx~=1.17.0
numpy~=1.26.0
pandas~=2.1.0
Pillow~=10.2.0
opencv-python~=4.9.0
